Type Check Elimination functionality added: If the field holding the current state is explicitly initialized with the value of a named constant, then it is initialized with an instance creation of the subclass corresponding to the specific named constant after the application of Replace Type Code with State/Strategy refactoring.
If the field holding the current state is implicitly
initialized (i.e. its field declaration does not have an
initializer), then it is initialized with an instance
creation of the subclass corresponding to the named
constant which has the zero value after the application of
Replace Type Code with State/Strategy refactoring.

git-svn-id: svn://java.uom.gr/jdeodorant@232 b7e68eb2-82c8-4b8b-a2e2-35572869c0fb