Removed condition "nodeId >= mappedNodes.first().getId() && nodeId <= mappedNodes.last().getId()" in the examination of precondition violations for unmapped statements.
In the case org.eclipse.emf.ecore.xml.type.internal.RegEx.matchesIgnoreCase
the first statement inside the do-while loop was unmatched.
However, its examination was skipped because of the aforementioned condition.
The do-while loop gets the ID following after the ID of the last statement inside its body.

The self-loop edges that are carried through a loop in the matched nodes
are also examined for precondition violations. 

git-svn-id: svn://java.uom.gr/jdeodorant/CloneRefactoring@636 b7e68eb2-82c8-4b8b-a2e2-35572869c0fb