Fixed CCFinder parser for Hibernate issue by making the paths absolute
The issue was making the parser not able to find any of the source
files, because it needed an extra input. We were explicitly giving this
input in the past, which was removed when we moved parsers to
JDeodorant. However, it turned out that we actually needed the input,
unless we used absolute paths. Now the problem has been fixed by using
absolute paths.