-Added support for postfix expressions in method modifyExtractedFieldAssignmentsInSourceClass()
-Changed the logic of method setPublicModifierToSourceMethod()
to update protected access modifier only for methods
in classes belonging to a different package than the
package of the source/extracted class.

-Extended Extract Class refactoring implementation
to update the access modifier of methods declared in the superclass
of the source class containing the methods to be extracted.

-Fixed the addition of qualifier for statically accessed source fields
in the extracted class (UNITS.length was excluded from being qualified).

-Fixed the naming of the getter methods
created for source fields accessed in the extracted class,
when a source method has the same name.

-Renaming the arguments passed to recursive extracted method invocations
for the non-extracted source fields accessed in the extracted methods
with "this" prefix

-Renaming the parameters passed to the extracted methods
for the non-extracted source fields accessed in the extracted methods
with "this" prefix according to
createNameForParameterizedFieldAccess()

-Passing fields as parameters to extracted methods
with a "this" field access
to avoid name collisions with local variables.