-Removed precondition containsReadObjectMethodAssigningExtractedField()
-Fixed method handleAccessedFieldNotHavingSetterMethod()
to replace properly the field accesses passed as parameters to the extracted methods
with a "this" prefix

-Extended Extract Class Refactoring
to update methods readObject() and writeObject()
so that they handle properly the reading and writing
of the field referring to the extracted class
from/to a stream.

-Extract Class candidates
containing one of the methods
readObject(java.io.ObjectInputStream)
writeObject(java.io.ObjectOutputStream)
are automatically excluded.